+++ src/makefile.linux-mingw
@@ -48,6 +48,7 @@ HEADERS = \
     main.h \
     net.h \
     noui.h \
+    protocol.h \
     rpc.h \
     script.h \
     serialize.h \
@@ -75,6 +76,7 @@ OBJS= \
     obj/keystore.o \
     obj/main.o \
     obj/net.o \
+    obj/protocol.o \
     obj/rpc.o \
     obj/script.o \
     obj/util.o \
