+++ src/makefile.vc
@@ -58,6 +58,7 @@ HEADERS = \
     main.h \
     net.h \
     noui.h \
+    protocol.h \
     rpc.h \
     script.h \
     serialize.h \
@@ -77,6 +78,7 @@ OBJS= \
     obj\main.o \
     obj\net.o \
     obj\rpc.o \
+    obj\protocol.o \
     obj\script.o \
     obj\util.o \
     obj\wallet.o \
